[WEB-72] Update paperwork page to diff rats
The paperwork page now builds a hash on submission. This hash contains a
list of rats that have been added to the rescue, as well as a list of
rats that have been removed from the rescue. The paperwork submission
will now handle assigning/unassigning those rats from the rescue.